requirements.txt
tensorflow==2.9.1
mediapipe==0.10.1
opencv-python==4.10.0.84
protobuf== 3.20.0
numpy==1.26.4
Pillow~=9.0.0
matplotlib~=3.9.2